### 2022-08-06 #4646 @elbeejay
@skiptoniam thank you for this submission to JOSS. Before asking any potential reviewers to have a look and decide if they are up to reviewing this submission, I have a number of comments and suggestions relating to the text of the manuscript. I encourage both you and your co-author to review the text of the manuscript after making any revisions before letting me know you are ready to proceed. To have the bot regenerate a PDF, write a comment that begins with `@editorialbot generate pdf`. My comments and suggestions are below and refer to line numbers in the PDF as generated above:

- Figure 1 runs off the page, please adjust its width using syntax like: `![Caption for example figure.](figure.png){ width=20% }`
- L 9: Do you mean to say "process models" rather than "processes models"?
- L 16: Should the "m" in "models" be capitalized as it is part of the IPPM acronym?
- L 17: I suggest changing to either "spatially referenced" or "spatial reference"
- L 26: Consider changing to read "package is setup" rather than "was" (similar suggestion for other instances referring to package design, as hopefully the software will continue to evolve and be maintained/developed)
- L 33: capitalize "Poisson"
- L 40: seems like "involved" should be "evolved"
- L 41: I'd suggest changing "process" to "processes" here
- Should it be "process" or "processes" in the IPPM acronym? Also is there a P missing in the acronym (Poisson Point Process)? The definition seems inconsistent between line 16 and the first subsection within Methods
- L 48: I found this sentence to be unclear, particularly the terminology "pattern in which n observed locations, say, makes over a spatial window" - is this a pattern made over some spatial window?
- L 57: should "point" be pluralized: "of the m quadrature points"?
- L 88: I would suggest deleting "which" after "rQuasi function"
- L 90: seems like a typo, "non" should be "not"?
- L 97: delete "While" or alter the structure of these sentences
- L 98: I would add words after "The reason for this" so that "this" is clear. For example: "The reason for this difference" - later in the sentence I think it either needs to be "due to the computational complexity of calculating" or "due to the computational complexities associated with calculating" or similar
- L 103: there is a period missing at the end of the sentence here
- L 117: "a" to "an" - I also found the next sentence to be unclear, especially the phrasing "under this scenario pass"
- L 119: consider changing "not distinguishing" to "not unique to"
- L 128: typo "the the"
- L 134, 136, 150: should be "an inhomogeneous" not "a"
- L 163: consider deleting "be" after "could"
- Overall, be consistent with the use of back-quotes ` to typeset references to code objects. For example sometimes ppmData is surrounded by back-quotes, other times it isn't.

Thank you,
Jay
