+++ paper/paper.md
@@ -39,19 +39,17 @@ editor_options:
 
 # Summary
 
-Spatial data, the locations of objects in a spatial region, is a common type of data recorded in ecology, environmental sciences and epidemiology amongst many others. An appropriate statistical approach for such data are spatial point processes models [e.g. @cressie_statistics_1993; @baddeley_practical_2000]. Spatial point processes can be used to understand the location of objects to inform decision making and improve ecological, health and other socio-economic outcomes. For example, the locations of individuals from a species' population are typically not distributed completely at random, but rather are observed at locations with favorable habitat conditions for that species. We might expect these species sightings are distributed according to relevant covariates.
+Spatial data, the locations of objects in a spatial region, is a common type of data recorded in ecology, environmental sciences and epidemiology amongst many others. An appropriate statistical approach for such data are spatial point process models [e.g. @cressie_statistics_1993; @baddeley_practical_2000]. Spatial point processes can be used to understand the location of objects to inform decision making and improve ecological, health and other socio-economic outcomes. For example, the locations of individuals from a species' population are typically not distributed completely at random, but rather are observed at locations with favorable habitat conditions for that species. We might expect these species sightings are distributed according to relevant covariates.
 
-Inhomogeneous Poisson Point Process models (IPPM) allow for the spatial intensity of a point process to vary across space, usually as a function of spatially reference covariates. One challenge when fitting an inhomogeneous Poisson point process is the integral used to described the intensity surface in the log-likelihood (as described below) has no known closed analytic solution. For IPPMs, the Berman-Turner device [@berman_approximating_1992] is commonly used to approximate the model using quadrature within a weighted Poisson generalized linear model. The type of quadrature scheme can have important ramifications for overall model fitting and inference [@warton_poisson_2010]. Approximating the integral accurately often comes at an increased computation cost, where a large number of quadrature points are required to robustly estimate the integral [@renner_point_2015].
+Inhomogeneous Poisson Process Models (IPPM) allow for the spatial intensity of a point process to vary across space, usually as a function of spatially referenced covariates. One challenge when fitting an Inhomogeneous Poisson Process Model is the integral used to described the intensity surface in the log-likelihood (as described below) has no known closed analytic solution. For IPPMs, the Berman-Turner device [@berman_approximating_1992] is commonly used to approximate the model using quadrature within a weighted Poisson generalized linear model. The type of quadrature scheme can have important ramifications for overall model fitting and inference [@warton_poisson_2010]. Approximating the integral accurately often comes at an increased computation cost, where a large number of quadrature points are required to robustly estimate the integral [@renner_point_2015].
 
[…]
