+++ paper/paper.Rmd
@@ -62,7 +62,7 @@ The approximate log-likelihood $\ell(\bm{\beta};\{\bm{x}_i\})$ can be written as
 ```
 where $z_i$ is an indicator variable identifying if site $i$ is one of the $n$ observed points from the random point pattern or one of the $m$ quadrature points. $w_i$ stores the quadrature weights which sum to the total area of the region $|\mathcal{A}|$. In \eqref{eq:two}, we can see the integral in the log-likelihood $\ell(\beta|\mathbf{y})$ is being estimated using quadrature [@baddeley_practical_2000; @diggle_geostatistical_2010]. @warton_poisson_2010 demonstrated for presence-only species distribution models, that using numerical quadrature with a point process framework makes the models scale invariant and treats the quadrature purely as a tool for approximating the integral.
 
-```{r fig1, fig.width=10, fig.height=5, fig.align='center',cache=TRUE, warning=FALSE, echo=FALSE, message=FALSE,fig.cap = "Quadrature schemes generated using the ppmData package for the species \\textit{Tasmaphena sinclairi} using Tasmania, Australia, as our spatial region. Red points are the known presences of \\textit{Tasmaphena sinclairi}. Black points are the background quadrature locations. A) Quasi-random quadrature where the integration points are generated using a quasi-random areal sample. Weights for each integration point are calculated as the dual of the subsequent Dirichlet tessellation of both the presence and integration points. B) Pseudo-random quadrature generated using random points from within the window. C) Grid quadrature where quadrature points are generated on a regular grid."}
+```{r fig1, fig.width=6, fig.height=3,fig.align='center',cache=TRUE, warning=FALSE, echo=FALSE, message=FALSE,fig.cap = "Quadrature schemes generated using the ppmData package for the species \\textit{Tasmaphena sinclairi} using Tasmania, Australia, as our spatial region. Red points are the known presences of \\textit{Tasmaphena sinclairi}. Black points are the background quadrature locations. A) Quasi-random quadrature where the integration points are generated using a quasi-random areal sample. Weights for each integration point are calculated as the dual of the subsequent Dirichlet tessellation of both the presence and integration points. B) Pseudo-random quadrature generated using random points from within the window. C) Grid quadrature where quadrature points are generated on a regular grid."}
 library(ppmData)
 path <- system.file("extdata", package = "ppmData")
 lst <- list.files(path=path, pattern='*.tif',full.names = TRUE)
