+++ paper/paper.Rmd
@@ -47,9 +47,9 @@ Our `ppmData` package is setup to use a quasi-random quadrature approach as an a
 
 Here we present the `ppmData` R package that is designed to setup a quadrature scheme using Dirichlet tessellation for fitting spatial point process models. Our initial intention was to develop multiple species (marked) quadrature in order to improve computation of multiple species Poisson point process models. Since then, the package has involved into one which can be useful for fitting univariate response inhomogeneous spatial point process, as discussed in this paper. We describe a how to set up a inhomogeneous Poisson point process quadrature and provide a simple demonstration on how one can fit an inhomogeneous Poisson point process using a `ppmData` object.
 
-# Inhomogeneous Poisson Process Models (IPPM)
+# Inhomogeneous Poisson Process Models
 
-An IPPM is a statistical model used to describe a random point pattern for $n$ observed locations, ($\bm{y}=(y(\bm{s}_1),y(\bm{s_2}),\ldots,y(\bm{s}_n))$, in a spatial window $\mathcal{A}\subset R^2$. An IPPM is based on the assumption that the number of points, $n$, is a Poisson random variable, with spatially varying intensity $\Lambda = \int_{\bm{s}\in\mathcal{A}}\lambda(\bm{s})d\bm{s}$, and that the spatial locations of $\bm{y}$ are independently drawn from the spatial surface [e.g. @cressie_statistics_1993]. For convenience, we define (for example) $\lambda(\bm{s}_i)$ as the intensity surface of the point process at the $i^{th}$ location $\bm{s}_i$. The intensity surface can be defined as $\log\{\lambda_i\}=\bm{x}_i^{\top}\bm{\beta}$ where $\bm{x}_i$ is a vector of spatially located covariates and $\bm{\beta}$ is a corresponding vector of coefficients.
+An Inhomogeneous Poisson Process Models is a statistical model used to describe a random point pattern for $n$ observed locations, ($\bm{y}=(y(\bm{s}_1),y(\bm{s_2}),\ldots,y(\bm{s}_n))$, in a spatial window $\mathcal{A}\subset R^2$. An Inhomogeneous Poisson Process Model assumes that the number of points, $n$, is a Poisson random variable, with spatially varying intensity $\Lambda = \int_{\bm{s}\in\mathcal{A}}\lambda(\bm{s})d\bm{s}$, and that the spatial locations of $\bm{y}$ are independently drawn from the spatial surface [e.g. @cressie_statistics_1993]. For convenience, we define (for example) $\lambda(\bm{s}_i)$ as the intensity surface of the point process at the $i^{th}$ location $\bm{s}_i$. The intensity surface can be defined as $\log\{\lambda_i\}=\bm{x}_i^{\top}\bm{\beta}$ where $\bm{x}_i$ is a vector of spatially located covariates and $\bm{\beta}$ is a corresponding vector of coefficients.
 
 The approximate log-likelihood $\ell(\bm{\beta};\{\bm{x}_i\})$ can be written as a weighted Poisson likelihood [@berman_approximating_1992]
 
