+++ paper/paper.Rmd
@@ -104,9 +104,9 @@ Here we use the quasi-random quadrature as a method to trade-off the robustness 
 
 The package `spatstat` also allows a user to develop a quasi-random quadrature scheme, but this appears to be quite slow when used with Dirichlet weights for larger numbers of dummy (quadrature) points (\> 10000). For example, a regular window with 100000 quadrature points takes approximately 3.5 minutes to run in `spatstat` using `quadscheme` with `method=Dirichlet`. Our package computes the same Dirichlet tessellation in about 4.5 seconds. Typically, calculating a Dirichlet tessellation increases super-linearly and can be come slow with for large numbers of points. Part of the challenge when developing `ppmData` was to make this step efficient by using C++ code to run a Delaunay triangulation via a radial sweep algorithm [@sinclair_s-hull_2016]. We then calculate the Dirichlet tessellations as the dual graph of this triangulation. Finally, areal weights for quadrature are calculated as the area of each polygon in the Dirichlet tessellation.
 
-## Generate a quadrature scheme using quasi-random sampling
+## Generating a quasi-random quadrature scheme
 
-In the `ppmData` package we have tried to make the generation of a quadrature scheme of a single or marked point process as easy as possible. We provide a simple interface for generating a Poisson point process data object and base all inputs and outputs using `terra` [@hijmans_terra_2021] and `sf` [@pebesma_simple_2018] packages. All we need to pass is a set of coordinates from a point process and we can generate a very basic point process data object. Ideally, the user knows the region they are interested in modelling and this is passed as a window, we use a `terra` SpatRaster object. The window defines the region where we expect to encounter observations and where to generate the quadrature locations. There are a number of other calls which can be passed to `ppmData`, but the main one is the inclusion of covariates which can be used in an Inhomogenenous Poisson Process Model. Under this scenario pass a set of covariates that share the same extent and resolution as the window SpatRaster. The values of the covariates will be extracted at the locations of the quadrature scheme. Whilst not unique to this package, the `ppmData` package can also generate grid and pseudo-random quadrature schemes as presented in Fig. 1.
+In the `ppmData` package we have tried to make the generation of a quadrature scheme of a single or marked point process as easy as possible. We provide a simple interface for generating a Poisson point process data object and base all inputs and outputs using `terra` [@hijmans_terra_2021] and `sf` [@pebesma_simple_2018] packages. Users need to pass a set of coordinates from an observed point pattern, a window to define the point pattern region and a set of covariates observed across this region. The window is defined using a `terra` SpatRaster object, and it used to identify the point pattern region and where to generate the quadrature locations. Covariates are included as a multiple layered `terra` SpatRaster object that share the same extent and resolution as the window. The values of the covariates will be extracted at the locations of the point pattern and quadrature scheme. Whilst not unique to this package, the `ppmData` package can also generate grid and pseudo-random quadrature schemes as presented in Fig. 1.
 
 ```{r multiple species quad, echo=FALSE, cache=TRUE, warning=FALSE, message=FALSE, fig.cap='Multiple species quadrature scheme. Here we can see the presences of five different snail species across Tasmania. For the multiple species quadrature there is a common set of background quadrature sites, but the object returns species specific weights.',fig.align='center',fig.height=5,fig.width=5}
 library(ppmData)
@@ -125,7 +125,7 @@ plot(sf::st_geometry(tassie),add=TRUE, border=gray(0.2,0.5))
 
 ## Multiple species (marked) point process quadrature
 
[…]
