### 2022-08-10 #4646 @skiptoniam
@elbeejay thanks so much for the suggestions and edits. I've tried to clean up those paper. I have also remove a bit of text including the future work section as the paper is fairly long. I was using rmarkdown to build the document, so hopefully the .md is OK now.
