### 2022-08-18 #4646 @elbeejay
@skiptoniam I still have a number of questions relating to the text of the manuscript, some of which are repeats of my previous comments that were not addressed.

- L 9-10: Shouldn't "e.g.," come before the references? Re-order this in the markdown file to read: `[e.g. @baddeley_practical_2000; @cressie_statistics_1993]` and this should be resolved
- L 18, 33-34, 46, 48: Is there a reason the IPPM acronym isn't being used in these instances?
- L 40: Should the word "involved" be "evolved" in this sentence?
- L 42: Should there be an "a" before "how" in this sentence? Should it be "an" rather than "a" preceding "inhomogeneous"?
- L 52: Why is "for example" in parentheses? Typically this would be written as a clause, e.g., "for particular spatial locations, for example, \Lambda_i ..."
- In the caption for Figure 1 is there a reason this "ppmData" is not typeset with backquotes?
- L 101: Is "be come" intentionally written this way? I'd have expected the single word "become"
- L 126: Is the repetition of the word "the" intentional in this sentence? This was a previous comment so I wanted to ask again, please let me know if this is intentional, maybe it is supposed to be "that the"?
- Figure 3: Is "recoreded" a typo? Maybe "recorded" is the intended word?

Again I'll encourage both you and your co-author to review the text of the manuscript after making any revisions. For the sake of clarity, please reply directly to any of the above questions if you are not planning to change the text in response to a given query. To have the bot regenerate a PDF, write a comment that begins with @editorialbot generate pdf.

Thanks!
