### 2022-09-01 #4646 @elbeejay
@skiptoniam following up here to ask if you have any outstanding questions or problems making these revisions to the text of the paper. Thanks!

Jay
