### 2022-09-08 #4646 @skiptoniam
Hi @elbeejay,

> @skiptoniam I still have a number of questions relating to the text of the manuscript, some of which are repeats of my previous comments that were not addressed.
>
> * L 9-10: Shouldn't "e.g.," come before the references? Re-order this in the markdown file to read: `[e.g. @baddeley_practical_2000; @cressie_statistics_1993]` and this should be resolved
_This has been resolved. I removed the e.g all together_

> * L 18, 33-34, 46, 48: Is there a reason the IPPM acronym isn't being used in these instances?
_I decided to remove the acronym for clarity. I'm not normally a fan of them, but it is quite a mouthful, so decided to use them originally. I have removed for now._

> * L 40: Should the word "involved" be "evolved" in this sentence?
_Fixed. Good catch._

> * L 42: Should there be an "a" before "how" in this sentence? Should it be "an" rather than "a" preceding "inhomogeneous"?
_I have fixed this sentence._

> * L 52: Why is "for example" in parentheses? Typically this would be written as a clause, e.g., "for particular spatial locations, for example, \Lambda_i ..."
_I'm not sure why I did that. The clause reads better :)_

> * In the caption for Figure 1 is there a reason this "ppmData" is not typeset with backquotes?
_Fixed with a bit of latex._

> * L 101: Is "be come" intentionally written this way? I'd have expected the single word "become"
_Typo. Fixed_

> * L 126: Is the repetition of the word "the" intentional in this sentence? This was a previous comment so I wanted to ask again, please let me know if this is intentional, maybe it is supposed to be "that the"?
_Apologies, that was an oversight by me_

> * Figure 3: Is "recoreded" a typo? Maybe "recorded" is the intended word?
_Typo. I have reworded this caption_

> Again I'll encourage both you and your co-author to review the text of the manuscript after making any revisions. For the sake of clarity, please reply directly to any of the above questions if you are not planning to change the text in response to a given query. To have the bot regenerate a PDF, write a comment that begins with @editorialbot generate pdf.
_We have looked over it again. Thank you for all your careful and considered edits, it is greatly appreciated._

> Thanks!

Once again apologies for the delay, I should have let you know I was away.

Regards,

Skip.
