+++ paper/paper.Rmd
@@ -49,7 +49,7 @@ Here we present the `ppmData` R package that is designed to setup a quadrature s
 
 # Inhomogeneous Poisson Process Models
 
-An Inhomogeneous Poisson Process Model is a statistical model used to describe a random point pattern for $n$ observed locations, ($\bm{y}=(y(\bm{s}_1),y(\bm{s}_2),\ldots,y(\bm{s}_n))$, in a spatial window $\mathcal{A}\subset R^2$. An Inhomogeneous Poisson Process Model assumes that the number of points in $\mathcal{A}$, $n$, is a Poisson random variable with total intensity $\Lambda$ defined as $\Lambda = \int_{\bm{s}\in\mathcal{A}}\lambda(\bm{s})d\bm{s}$. The spatial function $\lambda(\bm{s})$ is usually a non-constant function and defines the spatial surface that the spatial locations $\bm{y}$ are independently drawn from [@cressie_statistics_1993]. For convenience, we use subscript notation for particular spatial locations, for example, $\lambda_i=\lambda(\bm{s}_i)$ is the intensity surface of the point process at the $i^{th}$ location with coordinates $\bm{s}_i$. The intensity surface can be defined as $\log\{\lambda_i\}=\bm{x}_i^{\top}\bm{\beta}$ where $\bm{x}_i$ is a vector of spatially located covariates and $\bm{\beta}$ is a corresponding vector of coefficients.
+An Inhomogeneous Poisson Process Model is a statistical model used to describe a random point pattern for $n$ observed locations, ($\bm{y}=(y(\bm{s}_1),y(\bm{s}_2),\ldots,y(\bm{s}_n))$, in a spatial window $\mathcal{A}\subset R^2$. An Inhomogeneous Poisson Process Model assumes that the number of points in $\mathcal{A}$, $n$, is a Poisson random variable with total intensity $\Lambda$ defined as $\Lambda = \int_{\bm{s}\in\mathcal{A}}\lambda(\bm{s})d\bm{s}$. The spatial function $\lambda(\bm{s})$ is usually a non-constant function and defines the spatial surface that the spatial locations $\bm{y}$ are independently drawn from [@cressie_statistics_1993]. For convenience, we use subscript notation for particular spatial locations, for example, $\lambda_i=\lambda(\bm{s}_i)$ is the intensity surface of the point process at the $i^{th}$ location with coordinates $\bm{s}_i$. The intensity surface can be defined as $\log\{\lambda_i\}=\bm{x}_i^{\top} \bm{ \beta }$ where $\bm{x}_i$ is a vector of spatially located covariates and $\bm{\beta}$ is a corresponding vector of coefficients.
 
 The approximate log-likelihood $\ell(\bm{\beta};\{\bm{x}_i\})$ can be written as a weighted Poisson likelihood [@berman_approximating_1992]
 
