+++ paper/paper.Rmd
@@ -55,7 +55,7 @@ The approximate log-likelihood $\ell(\bm{{\beta}};\{\bm{x}_i\})$ can be written 
 
 ```{=tex}
 \begin{align}
-\ell(\bm{\beta}|\bm{y}) & = \sum_{i=1}^n\log(\lambda_i) - \int_{\bm{s} \in \mathcal{A}}\lambda(\bm{s})d\bm{s} - \log(n!) \nonumber\\
+\ell(\bm{{\beta}}|\bm{y}) & = \sum_{i=1}^n\log(\lambda_i) - \int_{\bm{s} \in \mathcal{A}}\lambda(\bm{s})d\bm{s} - \log(n!) \nonumber\\
 & \approx \sum_{i=1}^n\log(\lambda_i) - \sum_{i=1}^m w_i\lambda_i \nonumber\\
 & = \sum_{i=1}^m w_i\{z_i \log(\lambda_i) - \lambda_i\} \numberthis \label{eq:two}
 \end{align}
