+++ paper/paper.Rmd
@@ -134,7 +134,6 @@ Here we demonstrate how to fit an Inhomogeneous Poisson Process Model using `glm
 ```{r fit model, cache=TRUE, warning=FALSE, message=FALSE}
 #load relevant libraries
 library(ppmData)
-library(terra) ## need this load rasters
 #read the data into R
 path <- system.file("extdata", package = "ppmData")
 lst <- list.files(path=path,pattern='*.tif',full.names = TRUE)
