### 2022-09-19 #4771 @editorialbot
[REVIEW]: The ppmData R-package for setting up spatial point process models
**Submitting author:** @skiptoniam (<a href="http://orcid.org/0000-0001-7022-0509">Skipton Nicholas Charles Woolley</a>)
**Repository:** https://github.com/skiptoniam/ppmData
**Branch with paper.md** (empty if default branch): joss_submission
**Version:** v1.0.0
