### 2022-09-19 #4771 @elbeejay
@OwenWard and @mhesselbarth instructions for completing your reviews are provided in the top comment of this issue. Please make specific comments pertaining to your review in the ppmData repository itself, and reference this overarching JOSS review issue with a hyperlink, to maintain some connectivity between this review issue and individual repository issues. Feel free to provide a summary of your review here, and please ask any questions that you have about the process in this issue.

I will be asking the bot to provide reminder messages in 4 weeks. Right now we are requesting that our reviewers complete their reviews in 6 weeks if possible; if you need more time please just comment here so that both myself and @skiptoniam are aware.

Thanks!
Jay
