### 2022-09-21 skiptoniam/ppmData#17 @mhesselbarth
LICENSE
Regarding https://github.com/openjournals/joss-reviews/issues/4771

Please provide an [OSI approved](https://opensource.org/licenses/alphabetical) software license. In the `DESCRIPTION` it says `License: MIT + file LICENSE`, thus I would suggest to actually include a text file with the MIT license.
