### 2022-09-21 skiptoniam/ppmData#18 @mhesselbarth
Merge with ppmFit
Regarding https://github.com/openjournals/joss-reviews/issues/4771

Maybe it makes sense to merge the `ppmData` and `ppmFit` package into one package. I think the merged package would not be too big to successfully maintain and the functionality seems to be extremely related.

However, I also understand that there might be reasons to keep the packages separated. Thus, this suggestion could also be ignored by the authors with some reasoning.
