### 2022-09-21 skiptoniam/ppmData#19 @mhesselbarth
Check imports
Regarding https://github.com/openjournals/joss-reviews/issues/4771

It seems like some imports (i.e., `units`) are not used. Make sure to update the `DESCRIPTION` file.
