### 2022-09-21 skiptoniam/ppmData#20 @mhesselbarth
Replace class checks
Regarding https://github.com/openjournals/joss-reviews/issues/4771

Use `inherits()` instead of `if (class(x) == <CLASS>`
