### 2022-09-21 skiptoniam/ppmData#21 @mhesselbarth
testthat
Regarding https://github.com/openjournals/joss-reviews/issues/4771

It is appreciated and good practice that the package uses automated tests. However, currently none of the tests actually check if the function return values are correct using `testthat::expect_*()`. Thus, when running `evtools::test()` many of the tests are actually skipped because they are empty.
