### 2022-09-21 skiptoniam/ppmData#22 @mhesselbarth
Help file of ppmData
Regarding https://github.com/openjournals/joss-reviews/issues/4771

The help file of `?ppmData` uses `package file` as header, which it pulls from the `package.R` script. Please replace this with something more informative.

Related to this, maybe think about renaming your main function to a different name than the package name itself. However, this is more a personal suggestion from my side.
