### 2022-09-21 skiptoniam/ppmData#23 @mhesselbarth
External data
Regarding https://github.com/openjournals/joss-reviews/issues/4771

Maybe I am just missing this, but where is the external data coming from? Please document the source/reference somewhere.
