### 2022-09-21 skiptoniam/ppmData#24 @mhesselbarth
Methods to convert from and to spatstat
Regarding https://github.com/openjournals/joss-reviews/issues/4771

Given the large amount of functions related to point pattern by the `spatstat` library and its popularity, it might make sense to also provide convert-function for `spatstat` objects (e.g., `ppp`, `owin`, etc.)
