### 2022-09-21 skiptoniam/ppmData#25 @mhesselbarth
Code of Conduct
Regarding https://github.com/openjournals/joss-reviews/issues/4771

According to the JOSS criteria, each package needs community guidelines on how to contribute [(see here)](https://joss.readthedocs.io/en/latest/review_criteria.html). Please provide these information using e.g. a Code of Conduct file.
