### 2022-09-21 skiptoniam/ppmData#26 @mhesselbarth
Improve README
Regarding https://github.com/openjournals/joss-reviews/issues/4771

According to the JOSS criteria, the README file should contain information about how to install the package, as well as some guidance how to use it [(see here)](https://joss.readthedocs.io/en/latest/review_criteria.html). Please add those to the `README` file.
