### 2022-10-04 #4771 @mhesselbarth
> Would it be possible to include a vignette going through the example with the snail data in a bit more detail? I think this would be useful for people who may be unfamiliar with these types of problems.

Is there any recommendations what to put into the README and into vignettes in general? I opened https://github.com/skiptoniam/ppmData/issues/26 related to this, but personally don't have any strong opinion if that should be in the README or an additionally vignette. But some more guidance would be great!
