### 2022-10-04 #4771 @OwenWard
> Is there any recommendations what to put into the README and into vignettes in general? I opened [skiptoniam/ppmData#26](https://github.com/skiptoniam/ppmData/issues/26) related to this, but personally don't have any strong opinion if that should be in the README or an additionally vignette. But some more guidance would be great!

For JOSS I don't think so, but https://r-pkgs.org/vignettes.html has some ideas about vignettes.
My opinion would be that from the README you can tell if the package does what you want/need, and then if you are going to use it a vignette would give you the details to actually implement it.

I agree with the issue about the README, and would encourage (but not require) a vignette, so that it would hopefully lead to more people using the package.
