### 2022-10-04 #4771 @elbeejay
> For JOSS I don't think so, but https://r-pkgs.org/vignettes.html has some ideas about vignettes.

This is correct, as JOSS covers a wide array of programming languages we do not provide guidance on language-specific forms of documentation such as vignettes. Even for README files we provide only minimal suggestions as these can vary based on the language and community. Suggestions above are all appropriate as far as I can tell.
