+++ R/weights_grid.R
@@ -7,6 +7,7 @@
 #' @param window A SpatRaster from terra package which will represent the extent
 #'  and resolution of the point process model.
 #' @param coord The names of the coordinates. Default is c("X","Y").
+#' @param control A list of control options passed from ppmData
 #' @export
 #' @author Skipton Woolley
 #' @examples
@@ -21,7 +22,8 @@
 
 gridQuad <- function(npoints,
                      window,
-                     coord = c("X","Y")){
+                     coord = c("X","Y"),
+                     control){
 
   if(is.null(window)) stop("This function requires a window (terra raster) to work.")
   if(class(window)[1]!="SpatRaster") stop("'window' needs to be a 'SpatRaster' from the 'terra' package.")
@@ -37,7 +39,7 @@ gridQuad <- function(npoints,
   bk_sites <- terra::as.data.frame(bk_r,xy=TRUE)[,-3]
 
   nbk <- nrow(bk_sites)
-  message(sprintf("There are %s background sites generated using 'gridQuad' this is based on %s points, there will be less quadrature points than selected as na cells are removed and the grid is kept regular.", nbk,npoints))
+  if(!control$quiet)message(sprintf("There are %s background sites generated using 'gridQuad' this is based on %s points, there will be less quadrature points than selected as na cells are removed and the grid is kept regular.", nbk,npoints))
 
   bk_sites <- data.frame(bk_sites)
   colnames(bk_sites) <- coord
