+++ README.md
@@ -8,6 +8,8 @@ Status](https://codecov.io/github/skiptoniam/ppmData/coverage.svg?branch=master)
 
 ## ppmData is an R package that can be used to set up a quadrature scheme for spatial point processes modelling and could be used in [ecomix](https://github.com/skiptoniam/ecomix) to run multiple species Point Processes.
 
+### Summary
+
 The approach uses quasi-random sampling (Grafston & Tille, 2013, Foster
 et al., 2018) to generate a quadrature scheme for numerical
 approximation of a Poisson point process model (Berman & Turner 1992;
@@ -29,6 +31,49 @@ for generating pseudo-absences. If the users intent is to reduce
 sighting biases via a targeted background scheme, we recommend that bias
 is accounted for via covariates (distance to roads) or an offset (a
 known amount of effort) (e.g Warton et al., 2013; Renner et al, 2015).
+
+### Installation
+
+The `ppmData` package can be installed using `devtools` or `remotes` R
+packages.
+
+``` r
+install.packages('devtools')
+devtools::install_github('skiptoniam/ppmData')
+```
+
+### Example
+
+Here is an example using a dataset from the `ppmData` package. We setup
+a quadrature scheme for the species located within Tasmania, Australia.
+
+``` r
+library(ppmData)
+path <- system.file("extdata", package = "ppmData")
+lst <- list.files(path=path, pattern='*.tif',full.names = TRUE)
+preds <- rast(lst)
+presences <- subset(snails,SpeciesID %in% "Tasmaphena sinclairi")
+npoints <- 1000
+ppmdata1 <- ppmData(npoints = npoints, presences=presences,
+                    window = preds[[1]], covariates=preds)
+```
+
+Here we plot the quadrature scheme. The red points represent the known
+locations of . The black points represent the quadrature locations.
+Quasi-random quadrature where the integration points are generated using
+a quasi-random areal sample.
+
+``` r
+plot(ppmdata1)
+```
+
+<img src="README_files/figure-gfm/fig1-1.png" style="display: block; margin: auto;" />
+\## Code of Conduct
+
+Please note that the ppmData project is released with a [Contributor
+Code of
+Conduct](https://contributor-covenant.org/version/2/0/CODE_OF_CONDUCT.html).
+By contributing to this project, you agree to abide by its terms.
 
 ### References
 
