### 2022-10-14 86a29c5
update control for quad methods
