### 2022-10-18 #4771 @OwenWard
@elbeejay It seems like there have been some recent updates on the JOSS branch for this repo, so maybe it would make sense for @mhesselbarth and I to wait until an update that they have been completed before we finish the review? They seem to be addressing some of the previously mentioned comments.
