### 2022-10-23 #4771 @mhesselbarth
Some comments considering the manuscript. Overall its written very well and does a good job explaining the package and method. Just some minor suggestions and comments:

- Line 8/9: This is an appropriate statistical approach only if locations can be meaningful described by points (e.g., as for plants, nests).

- Line 13/14: This is one common driver, but also other processes can be more important for the pattern (e.g., competition, neutral theory, etc.). Thus, I would suggest to reword using e.g., “such as”, “for example”, “besides others”.

- Line 15: I would suggest to use locations instead of sightings to be concise.

- Line 16: Maybe add one more sentence to the previous section to better link habitat conditions to inhomogeneous poisson process models. That should help the reader better to connect the two.

- Line 17-19: I think there is a "that" missing in the sentence.

- Line 36-42: Please reword and state only what the package is intended and capable of doing now. Personally, I think the initial intention part is a bit confusing. Also, as far as I understand it, the package can be used for any qualitative marks and not just species. Thus, I would reword the sentence and using species as one example of qualitative marks.

- Line 66: I understand why the Statement of Need is introduced so late because the reader needs some background about the point processes first. However, I think it might also make sense to move this part up, or at least refer to it earlier in the introduction. But if you decided to keep the structure as it is right now that would be also fine for me.

- Line 95-99: Please add some information on what kind of computer/laptop this benchmark is based on (number of cores, RAM, operating system, etc.).

- Line 107-109: As also mentioned in issue #24. I would strongly suggest to add conversion methods for spatstat objects or at least explain somewhere in the Vignettes how to convert between spatstat and terra/sf given how popular the spatstat package is.

- Line 140/141: This sentence could use some more explanation. Why is the step required?
