### 2022-10-23 skiptoniam/ppmData#29 @mhesselbarth
URL and BugReports
Regarding https://github.com/openjournals/joss-reviews/issues/4771

Please add a general and bug report URL to the DESCRIPTION file (`URL` and `BugReports` fields).
