### 2022-11-01 #4771 @OwenWard
Just wanted to follow up also with a bit more detail on my comments. I don't think there is any large enough for an issue which @mhesselbarth hasn't already covered but I'll list my comments again. These include the current updates in the `joss_submission` branch.

- Can you provide a reference for the external data? It might be easier to just include it in the package directly? then it would get a help file, etc
- There isn’t an appropriate license for JOSS included
- The tests which are included don’t currently test any of the methods of the package.
- There should be a section detailing how to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
- Given that the statement of need for this paper emphasizes the computational benefits, I think an experiment demonstrating the speedup should be included in the JOSS paper, even commented out, so that readers can confirm it for themselves. I think you could also demonstrate this in the Readme.
- I'm not sure of the role of the file `package.R`? It also seems to be somehow linked to the documentation and comes up when you run `?ppmData`
- There are some small typos in the ReadMe in the JOSS branch
    - In the summary, “Quasi random sampling quadrature are form” and “thus reducing low errors”
    - In the example the text mentions red points which are now green
- Also some small typos in the paper on top of those mentioned previously, nothing much
    - Page 3 line 77, extra “is”
    - Page 4 line 111-112 maybe missing “is”?
    - Page 5 line 130 using “the” glm function maybe?
    - The second reference in the bibliography has some missing latex

Otherwise, when all these points are addressed I'd like to take another quick look at the code.
