### 2022-11-02 fc5222f
add na.rm for covariates/single species ppm
