+++ NAMESPACE
@@ -4,6 +4,7 @@ S3method(plot,dirTess)
 S3method(plot,ppmData)
 S3method(polygonise,dirTess)
 S3method(print,ppmData)
+export(as.spatstat)
 export(dirTess)
 export(gridQuad)
 export(gridWeights)
@@ -12,6 +13,7 @@ export(ppmData)
 export(pseudoRandomQuad)
 export(pseudoRandomWeights)
 export(rast_to_sf)
+export(terra2im)
 importFrom(Rcpp,sourceCpp)
 importFrom(grDevices,gray)
 importFrom(graphics,legend)
@@ -29,6 +31,8 @@ importFrom(sf,st_crs)
 importFrom(sf,st_intersection)
 importFrom(sf,st_make_valid)
 importFrom(sf,st_sf)
+importFrom(spatstat.geom,im)
+importFrom(spatstat.geom,transmat)
 importFrom(stats,complete.cases)
 importFrom(stats,median)
 importFrom(stats,runif)
@@ -38,6 +42,7 @@ importFrom(terra,ext)
 importFrom(terra,extract)
 importFrom(terra,freq)
 importFrom(terra,global)
+importFrom(terra,hasValues)
 importFrom(terra,ifel)
 importFrom(terra,init)
 importFrom(terra,mask)
