### 2022-11-23 68afb44
add spatstat utilty.
