+++ man/as.spatstat.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/spatstat_utils.R
+\name{as.spatstat}
+\alias{as.spatstat}
+\title{Function to covert a spatstat quadrature}
+\usage{
+as.spatstat(object, ...)
+}
+\arguments{
+\item{object}{A ppmData object}
+
+\item{\\dots}{Ignored}
+}
+\description{
+Function to covert a spatstat quadrature
+}
+\examples{
+\dontrun{
+library(ppmData)
+path <- system.file("extdata", package = "ppmData")
+lst <- list.files(path=path,pattern='*.tif',full.names = TRUE)
+preds <- rast(lst)
+window <- preds[[1]]
+presences <- subset(snails,SpeciesID \%in\% "Tasmaphena sinclairi")
+object <- ppmData(npoints = 10000, presences=presences, window = window, covariates = preds)
+ss.object <- as.spatstat(object)
+ppmfit <- ppm(ss.object$Q~1+poly(min_temp_coldest_month,2)+poly(annual_mean_precip,2),data=ss.object$covariates,Poisson())
+pred <- predict(ppmfit, covariates = ss.object$covariates,type="trend")
+}
+}
