### 2022-11-23 5683f16
added spatstat as.spatstat as.ppp & as.owin
