+++ man/as.owin.ppmData.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/spatstat_utils.R
+\name{as.owin.ppmData}
+\alias{as.owin.ppmData}
+\title{Convert a ppmData pnkect to an owin from the SpatStat package.}
+\usage{
+\method{as.owin}{ppmData}(W, ..., fatal = TRUE)
+}
+\arguments{
+\item{W}{Window.}
+
+\item{...}{Unused.}
+
+\item{fatal}{Unused.}
+}
+\description{
+Convert a ppmData pnkect to an owin from the SpatStat package.
+}
