+++ man/as.ppp.Rd
@@ -1,0 +1,31 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/spatstat_utils.R
+\name{as.ppp}
+\alias{as.ppp}
+\alias{as.ppp.ppmData}
+\title{Function to covert a spatstat quadrature}
+\usage{
+as.ppp.ppmData(X, ..., fatal = TRUE)
+}
+\arguments{
+\item{X}{A ppmData object}
+
+\item{fatal}{Logical value specifying what to do if the data cannot be converted. See Details.}
+
+\item{\\dots}{Ignored}
+}
+\description{
+Function to covert a spatstat quadrature
+}
+\examples{
+\dontrun{
+library(ppmData)
+path <- system.file("extdata", package = "ppmData")
+lst <- list.files(path=path,pattern='*.tif',full.names = TRUE)
+preds <- rast(lst)
+window <- preds[[1]]
+presences <- subset(snails,SpeciesID \%in\% "Tasmaphena sinclairi")
+object <- ppmData(npoints = 10000, presences=presences, window = window, covariates = preds)
+X <- as.ppp(object)
+}
+}
