+++ NAMESPACE
@@ -1,9 +1,11 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(as.owin,ppmData)
 S3method(plot,dirTess)
 S3method(plot,ppmData)
 S3method(polygonise,dirTess)
 S3method(print,ppmData)
+export(as.ppp.ppmData)
 export(as.spatstat)
 export(dirTess)
 export(gridQuad)
@@ -31,7 +33,11 @@ importFrom(sf,st_crs)
 importFrom(sf,st_intersection)
 importFrom(sf,st_make_valid)
 importFrom(sf,st_sf)
+importFrom(spatstat.geom,as.owin)
+importFrom(spatstat.geom,as.ppp)
 importFrom(spatstat.geom,im)
+importFrom(spatstat.geom,ppp)
+importFrom(spatstat.geom,quad)
 importFrom(spatstat.geom,transmat)
 importFrom(stats,complete.cases)
 importFrom(stats,median)
