### 2022-11-23 skiptoniam/qrbp#24 @skiptoniam
added as.spatstat, as.ppp and as.owin for some spatstat integration.
