### 2022-11-23 skiptoniam/qrbp#24 @mhesselbarth
Cool, looking good to me. For the future, would be easier if you tag the commits in the issue. Alternatively you can include "fix #<issue>" in your commit and automatically close the issue by that.

https://github.com/skiptoniam/ppmData/commit/5683f1607e22f179712f001b7f93e88deab9a2d4
