+++ DESCRIPTION
@@ -9,6 +9,8 @@ Description: This package generates a set of quasi-random points for a quadratur
 License: MIT + file LICENSE
 LazyData: TRUE
 RoxygenNote: 7.2.0
+URL: https://github.com/skiptoniam/ppmData
+BugReports: https://github.com/skiptoniam/ppmData/issues
 Imports: 
     methods,
     mgcv,
