+++ NAMESPACE
@@ -1,11 +1,11 @@
 # Generated by roxygen2: do not edit by hand
 
 S3method(as.owin,ppmData)
+S3method(as.ppp,ppmData)
 S3method(plot,dirTess)
 S3method(plot,ppmData)
 S3method(polygonise,dirTess)
 S3method(print,ppmData)
-export(as.ppp.ppmData)
 export(as.spatstat)
 export(dirTess)
 export(gridQuad)
