+++ DESCRIPTION
@@ -2,9 +2,9 @@ Package: ppmData
 Type: Package
 Title: Quasirandom quardature scheme using dirichlet tesselation for spatial Poisson point process models
 Version: 1.0.0
-Authors@R: c(person("Skipton", "Woolley", email = "skipton.woolley@unimelb.edu.au", role = c("aut", "cre")),
-             person("Scott", "Foster", email = "scott.foster@csiro.au", role = c("aut")))
-Maintainer: Skipton Woolley <skipton.woolley@unimelb.edu.au>
+Authors@R: c(person("Skipton", "Woolley", email = "skip.woolley@csiro.au", role = c("aut", "cre")),
+             person("Scott", "Foster", email = "scott.foster@data61.csiro.au", role = c("aut")))
+Maintainer: Skipton Woolley <skip.woolley@csiro.au>
 Description: This package generates a set of quasi-random points for a quadrature scheme, we calculate each points weights using a dirichlet tesselation (voronoi) algorithm. The resulting object can be used in existing single and multiple species models in order to extend them to Point Processes.
 License: MIT + file LICENSE
 LazyData: TRUE
@@ -18,7 +18,6 @@ Imports:
     sf,
     randtoolbox,
     Rcpp,
-    units,
     spatstat.geom
 Suggests: 
     knitr,
