### 2022-11-25 skiptoniam/qrbp#20 @skiptoniam
I have updated this - I used a mix of `inherits` and `isa`
