### 2022-12-01 d9b8fec
Update the package info to enable a bit more info "fix #22"
