+++ R/methods.R
@@ -41,13 +41,13 @@ plot.ppmData <- function(x,
       terra::plot(p,col=rcols,axes=FALSE, legend=FALSE, main=main)
     }
     points(quad,pch='.')
-    points(marks,col=as.character(colshp[as.numeric(as.factor(marks[,x$params$species.id])),1]),
-           pch=colshp[as.numeric(as.factor(marks[,x$params$species.id])),2], ...)
+    points(marks,col=as.character(colshp[as.numeric(as.factor(marks[,x$params$mark.id])),1]),
+           pch=colshp[as.numeric(as.factor(marks[,x$params$mark.id])),2], ...)
     if(plt.legend){
     legend(x="bottom",
-           legend=unique(marks[,x$params$species.id]),
-           col=as.character(colshp[1:length(unique(marks[,x$params$species.id])),1]),
-           pch=colshp[1:length(unique(marks[,x$params$species.id])),2],
+           legend=unique(marks[,x$params$mark.id]),
+           col=as.character(colshp[1:length(unique(marks[,x$params$mark.id])),1]),
+           pch=colshp[1:length(unique(marks[,x$params$mark.id])),2],
            cex=0.75,
            xpd = TRUE, horiz = FALSE, inset = c(-.2, 0),
            bty="n")
@@ -71,7 +71,7 @@ plot.ppmData <- function(x,
     points(pressies,col=pch.cols[1],pch=16, ...)
     if(plt.legend){
     legend(x="bottom",
-           legend=unique(pressies[,x$params$species.id]),
+           legend=unique(pressies[,x$params$mark.id]),
            col=pch.cols[1],
            pch=16,
            cex=0.75,
@@ -93,7 +93,7 @@ print.ppmData <- function (x, ...){
 
   if(!x$marked){
 
-    message("There are ", sum(x$ppmData$presence), " presence observations for this species")
+    message("There are ", sum(x$ppmData$presence), " presence observations for this point process")
     message("There are ", sum(x$ppmData$presence==0), " background quadrature (integration) points")
     message("There are a total of ", nrow(x$ppmData), " sites in the model.matrix")
     no_nans <- nrow(x$ppmData[!complete.cases(x$ppmData),])
@@ -102,8 +102,8 @@ print.ppmData <- function (x, ...){
 
   } else {
 
-    message("There are ", x$ppmData$nUniquePres, " presence observations for ", x$ppmData$nspp," species")
-    message("There are ", x$ppmData$nBkg, " quadrature (integration) points for each of the ", x$ppmData$nspp ," species")
+    message("There are ", x$ppmData$nUniquePres, " presence observations for ", x$ppmData$nmark," species")
+    message("There are ", x$ppmData$nBkg, " quadrature (integration) points for each of the ", x$ppmData$nmark ," species")
     message("There are a total of ",x$ppmData$m, " sites in the model.matrix")
     if(!is.null(x$ppmData$covars)){
       num_nans <- nrow(x$ppmData$covars[!complete.cases(x$ppmData$covars),])
