+++ R/weights_grid.R
@@ -26,7 +26,7 @@ gridQuad <- function(npoints,
                      control){
 
   if(is.null(window)) stop("This function requires a window (terra raster) to work.")
-  if(class(window)[1]!="SpatRaster") stop("'window' needs to be a 'SpatRaster' from the 'terra' package.")
+  if(!inherits(window,"SpatRaster")) stop("'window' needs to be a 'SpatRaster' from the 'terra' package.")
 
   ## fix this to regular grid
   bk_r <- terra::spatSample(x = window,
@@ -59,7 +59,7 @@ gridQuad <- function(npoints,
 #' @param window A SpatRaster from terra package which will represent the extent
 #'  and resolution of the point process model.
 #' @param coord The names of the coordinates. Default is c("X","Y").
-#' @param species.id character Name of the species ID column
+#' @param mark.id character Name of the species ID column
 #' @param unit The scale of the area weights, default is kilometers squared "km"
 #' but meters squared "m" or hectars "ha" can be used.
 #' @author Skipton Woolley
@@ -82,7 +82,7 @@ gridWeights <- function(presences,
                         quadrature,
                         window,
                         coord = c("X","Y"),
-                        species.id,
+                        mark.id,
                         unit = c("geo","m","km","ha")){
 
   # work out what unit you want the areas in.
@@ -92,7 +92,7 @@ gridWeights <- function(presences,
   if(is.null(quadrature)) stop("This function needs a set of quadrature sites to work")
 
   if(is.null(window)) stop("This function requires a window (terra raster) to work.")
-  if(class(window)[1]!="SpatRaster") stop("'window' needs to be a 'SpatRaster' from the 'terra' package.")
+  if(!inherits(window,"SpatRaster")) stop("'window' needs to be a 'SpatRaster' from the 'terra' package.")
 
   ## expanse broke for large raster
   npres <- nrow(presences)
@@ -104,21 +104,9 @@ gridWeights <- function(presences,
 
   ## Tracking of site and species ids
   allpts$id <- seq_len(nrow(allpts))
-  allpts$dataset <- allpts.id[,species.id]
+  allpts$dataset <- allpts.id[,mark.id]
 
   # if in geographic coordinates return area in unit of interest.
-  # if(terra::is.lonlat(window)){
-    # if(unit=="geo"){ # useful if you just want areas on the geographic scale (this is what deldir gives you)
-    #   areas <- terra::init(window,prod(terra::res(window)))
-    #   bck_wts <- terra::extract(areas,allpts[,coord])[,2] # drop ID
-    #   bck_wts_num <- list2numeric(bck_wts)
-    # } else {
-      # areas <- terra::cellSize(window, unit = unit)
-       # bck_wts <- terra::extract(areas,allpts[,coord],method="bilinear")[,2]
-    #   bck_wts_num <- list2numeric(bck_wts)
-    # }
-
-  # }else{
     if(unit=="geo"){
       areas <- terra::init(window,prod(terra::res(window)))
       areas.m <- terra::mask(areas,window)
@@ -129,19 +117,10 @@ gridWeights <- function(presences,
       bck_wts_site <- terra::extract(areas,allpts[allpts$dataset=="quad",coord],method="bilinear")[,2]
       window_area <- as.numeric(terra::global(areas, "sum", na.rm=TRUE))
       window_wts <- bck_wts_site/sum(bck_wts_site,na.rm = TRUE)
-      # sum(window_wts*window_area) # sanity check
       bck_wts  <- window_wts*window_area
     }
-  # }
-
[…]
