+++ R/weights_pseudo.R
@@ -24,7 +24,7 @@ pseudoRandomQuad <- function(npoints,
                              control){
 
   if(is.null(window)) stop("This function requires a window (terra raster) to work.")
-  if(class(window)[1]!="SpatRaster") stop("'window' needs to be a 'SpatRaster' from the 'terra' package.")
+  if(!inherits(window,"SpatRaster")) stop("'window' needs to be a 'SpatRaster' from the 'terra' package.")
 
   background_sites <- terra::spatSample(x = window,
                                         size = npoints,
@@ -51,7 +51,7 @@ pseudoRandomQuad <- function(npoints,
 #' @param window A SpatRaster from terra package which will represent the extent
 #'  and resolution of the point process model.
 #' @param coord The names of the coordinates. Default is c("X","Y").
-#' @param species.id character Name of the species ID column
+#' @param mark.id character Name of the species ID column
 #' @param unit The scale of the area weights, default is kilometers squared "km"
 #' but meters squared "m" or hectars "ha" can be used.
 #' @author Skipton Woolley
@@ -74,7 +74,7 @@ pseudoRandomWeights <- function(presences,
                                 quadrature,
                                 window,
                                 coord,
-                                species.id,
+                                mark.id,
                                 unit = c("geo","m","km","ha")){
 
   # work out what unit you want the areas in.
@@ -84,7 +84,7 @@ pseudoRandomWeights <- function(presences,
   if(is.null(quadrature)) stop("This function needs a set of quadrature sites to work")
 
   if(is.null(window)) stop("This function requires a window (terra raster) to work.")
-  if(class(window)[1]!="SpatRaster") stop("'window' needs to be a 'SpatRaster' from the 'terra' package.")
+  if(!inherits(window,"SpatRaster")) stop("'window' needs to be a 'SpatRaster' from the 'terra' package.")
 
   ## expanse broke for large raster
   npres <- nrow(presences)
@@ -96,7 +96,7 @@ pseudoRandomWeights <- function(presences,
 
   ## Tracking of site and species ids
   allpts$id <- seq_len(nrow(allpts))
-  allpts$dataset <- allpts.id[,species.id]
+  allpts$dataset <- allpts.id[,mark.id]
 
   if(unit=="geo"){
     areas <- terra::init(window,prod(terra::res(window)))
@@ -108,7 +108,6 @@ pseudoRandomWeights <- function(presences,
     bck_wts_site <- terra::extract(areas,allpts[allpts$dataset=="quad",coord],method="bilinear")[,2]
     window_area <- as.numeric(terra::global(areas, "sum", na.rm=TRUE))
     window_wts <- bck_wts_site/sum(bck_wts_site,na.rm = TRUE)
-    # sum(window_wts*window_area) # sanity check
     bck_wts  <- window_wts*window_area
   }
 

[…]
