+++ README.Rmd
@@ -28,7 +28,7 @@ devtools::install_github('skiptoniam/ppmData')
 
 ### Example
 
-Here is an example using a dataset from the `ppmData` package. We setup a quadrature scheme for the species \textit{Tasmaphena sinclairi} located within Tasmania, Australia. 
+Here is an example using a dataset from the `ppmData` package. We setup a quadrature scheme for the species *Tasmaphena sinclairi* located within Tasmania, Australia. 
 
 ```{r, message=FALSE}
 library(ppmData)
@@ -41,11 +41,13 @@ ppmdata1 <- ppmData(npoints = npoints, presences=presences,
                     window = preds[[1]], covariates=preds)
 ```
 
-Here we plot the quadrature scheme. The red points represent the known locations of \textit{Tasmaphena sinclairi}. The black points represent the quadrature locations. Quasi-random quadrature where the integration points are generated using a quasi-random areal sample.
+Here we plot the quadrature scheme. The green points represent the known locations of *Tasmaphena sinclairi*. The black points represent the quadrature locations. Quasi-random quadrature where the integration points are generated using a quasi-random areal sample.
 
 ```{r fig1, fig.width=7, fig.height=4, fig.align='center', echo=TRUE}
 plot(ppmdata1)
 ```
+Once the `ppmData` object has been set up it is quiet easy to fit a point process model in R. You can use something like `glm`, `gam` or even convert the `ppmData` object to work directly with `spatstat::ppm`.
+
 ## Code of Conduct
   
 Please note that the ppmData project is released with a [Contributor Code of Conduct](https://contributor-covenant.org/version/2/0/CODE_OF_CONDUCT.html). By contributing to this project, you agree to abide by its terms.
