### 2022-12-01 af60c51
added description and references for the raster datasets. "fix #23" & "close #23"
