### 2022-12-02 #4771 @elbeejay
@skiptoniam just wanted to check-in here and find out how the revisions are going. It looks like you've made a number of commits to the `joss_submission` branch so that is encouraging. Do you have a revised timeline for when you think you'll be ready to ask our reviewers to have another look?
