+++ tests/testthat/test_single_spp_ppmdata.R
@@ -15,44 +15,71 @@ npoints <- 1000
 ## some test a-roos!
 # test with just presences
 ppmdata0 <- ppmData(npoints = npoints, presences = presences)
-plot(ppmdata0)
+testthat::expect_s3_class(ppmdata0,"ppmData")
 
+## error if no presences
+testthat::expect_error(ppmData(npoints = npoints))
+
+## error if presences in the wrong format message to tell correct format.
+testthat::expect_error(ppmData(presences = "a",npoints = npoints))
 
 # test with just pres & window
 ppmdata1 <- ppmData(presences=presences, window = preds[[1]], npoints = npoints)
-plot(ppmdata1)
+testthat::expect_s3_class(ppmdata1,"ppmData")
+
+## test with a dodgy window
+testthat::expect_error(ppmData(presences=presences, window = "window", npoints = npoints))
+## show throw an error if a bounding box is given rather than a spatial raster
+win <- ppmData:::get_bbox(presences[,1:2])
+testthat::expect_error(ppmData(presences=presences, window = win, npoints = npoints))
 
 # test with just pres, window & covars
 ppmdata2 <- ppmData(presences=presences, window = preds[[1]], covariates = preds, npoints = npoints)
-plot(ppmdata2)
+testthat::expect_s3_class(ppmdata2,"ppmData")
+testthat::expect_error(ppmData(presences=presences, window = preds[[1]], covariates = as.data.frame(preds), npoints = npoints))
+
 
 # test with consistant masks.
 x <- sum(preds)
 preds2 <- mask(preds,x)
 ppmdata2a <- ppmData(npoints = npoints, presences=presences, window = preds2[[1]], covariates = preds2)
+testthat::expect_s3_class(ppmdata2a,"ppmData")
 
 # test with no npoints
 ppmdata3 <- ppmData(presences=presences, window = preds[[1]])
-plot(ppmdata3)
+testthat::expect_s3_class(ppmdata3,"ppmData")
+
 
 ppmdata4 <- ppmData(npoints=1000,
                     presences=presences,
                     window = preds[[1]],
                     unit="m")
+testthat::expect_s3_class(ppmdata4,"ppmData")
 
 ppmdata5 <- ppmData(npoints=1000,
                     presences=presences,
                     window = preds[[1]],
                     unit="km")
+testthat::expect_s3_class(ppmdata5,"ppmData")
 
 ppmdata6 <- ppmData(npoints=1000,
                     presences=presences,
                     window = preds[[1]],
                     unit="ha")
+testthat::expect_s3_class(ppmdata6,"ppmData")
 
-sum(ppmdata4$ppmData$weights)
-sum(ppmdata5$ppmData$weights)
-sum(ppmdata6$ppmData$weights)
+## check areal sums
+testthat::expect_gt(sum(ppmdata4$ppmData$weights),6e10)
+testthat::expect_gt(sum(ppmdata5$ppmData$weights),6e4)
+testthat::expect_gt(sum(ppmdata6$ppmData$weights),6e6)
+
+## throw some errors if wrong unit
+testthat::expect_error(ppmData(npoints=1000,
+                               presences=presences,
+                               window = preds[[1]],
+                               unit="meters"))
+
+
 
 })
 
@@ -68,56 +95,89 @@ testthat::test_that('ppmdata single species random', {
 
   ## some test a-roos!
   # test with just presences
-  ppmdata0 <- ppmData(npoints = npoints, presences = presences, quad.method = 'pseudo.random')
-  plot(ppmdata0)
+  ppmdata0 <- ppmData(npoints = npoints,
+                      presences = presences,
+                      quad.method="pseudo.random")
+  testthat::expect_s3_class(ppmdata0,"ppmData")
 
+  ## error if no presences
[…]
