+++ tests/testthat/test_multiple_spp_ppmdata.R
@@ -1,5 +1,3 @@
-context('ppmData multiple species')
-
 testthat::test_that('ppmData multiple species', {
 
 library(ppmData)
