### 2022-12-06 #4771 @skiptoniam
@elbeejay I'm hoping to finish up the tests and a few other things, before cleaning up the manuscript, hopefully I'll be done by the end of this week.
