### 2022-12-07 5ebd511
add spatstat utils checks
