### 2022-12-07 1f3da04
update spatstat tests
