+++ .github/workflows/test-coverage.yaml
@@ -2,9 +2,9 @@
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    branches: [main, master]
+    branches: [main, master, joss_submission]
   pull_request:
-    branches: [main, master]
+    branches: [main, master, joss_submission]
 
 name: test-coverage
 
