### 2022-12-13 skiptoniam/qrbp#18 @skiptoniam
At the moment, ppmFit is a fair way off being ready. And there are some methods still being worked up for this package - which might end up in other manuscripts. I do think this is a good idea in the long run. But I won't address if for this review https://github.com/openjournals/joss-reviews/issues/4771. I will close this issue for now.
