### 2022-12-13 skiptoniam/qrbp#21 @skiptoniam
I have updated the tests - I normally just check for code coverage, but this is a much better way of doing things. Thanks for the heads up.
