### 2022-12-14 0aa519b
add point process vignette
