+++ vignettes/ppmData_tutorial.Rmd
@@ -1,0 +1,233 @@
+---
+title: "How to use ppmData to fit a Inhomogenous Poisson Point Process."
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{How to use ppmData to model a species' distribution.}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+Load the required libraries
+```{r setup}
+library(ppmData)
+```
+
+## Setting up a ppm quadrature.
+Basically you need a set of environmental or spatial rasters which will be used 
+as covariates in the inhomogeneous point process models to estimate the intensity 
+of the point pattern. In this case we are using a snail species from Tasmania, 
+Australia. We aim to use `ppmData` to set up a quadrature scheme which can be 
+used in any glm style model to fit a weighted Poisson GLM. For example, users could
+use `mgcv` or `glmnet` for extra flexibility of regression coefficients or 
+regularization of covariates, respectively. 
+
+`ppmData` comes with a set of species presences and raster data to run examples 
+in the package and vignettes. There is a marked point pattern for the recorded 
+locations of a group of terrestrial snails observed in Tasmania, Australia. The 
+species identification represent a mark at each location. There is also a set of
+covariates that include some commonly used environmental gradients and a few 
+observer layers (such as distance from main roads).
+
+For this example we will just use a single species, and treat it a s regular
+point pattern. 
+```{r, cache=TRUE}
+path <- system.file("extdata", package = "ppmData")
+lst <- list.files(path=path,pattern='*.tif',full.names = TRUE)
+covariates <- rast(lst)
+presences <- subset(snails,SpeciesID %in% "Tasmaphena sinclairi")
+presences <- presences[!duplicated(presences),]
+```
+
+We can plot the recorded presences of *Tasmaphena sinclairi*. 
+```{r, fig.width=6, fig.height=6, fig.cap="Recorded presences of *Tasmaphena sinclairi* in Tasmania, Australia.", cache=TRUE }
+plot(covariates[[1]], legend=FALSE,axes=FALSE)
+points(presences,pch=16,cex=0.75,col=gray(0.2,0.75))
+```
+
+To fit a inhomogeneous Poisson point process for *Tasmaphena sinclairi* we need to set up a
+quadrature scheme. We can do this using the ppmData function. We need to pass a 
+set of presences, a raster which represents a window or region in which the 
+presences are observed and the model is to be fitted. We also need to set  of covariates 
+which we'll use to describe the distribution of the observed presences.
+```{r, cache=TRUE}
+window <- covariates[[1]]
+ppmdata <- ppmData(presences=presences,
+                   window = window,
+                   covariates = covariates)
+```
+We can now plot the quadrature scheme and see what this looks like within our 
+window.
+```{r, fig.width=6, fig.height=6, fig.cap="Quadrature scheme for *Tasmaphena sinclairi*.", cache=TRUE}
+plot(ppmdata)
+```
+
+In the example above, we have not provided the number of quadrature (background)
+points for modelling and really need to do so if we want to generate a 
+reasonable model for point process, there is actually a trade off between the 
+number of quadrature points and the computation time required to run a ppm. As 
+the number of quadrature points ($Q$) approaches infinity ($Q_{n\to\infty}$) the
+integrate estimated using the quadrature should become more and more accurate, 
+at the cost of greater computation time. 
+
[…]
