+++ paper/paper.bib
@@ -19,9 +19,8 @@
 	issn = {1467-842X},
 	url = {http://onlinelibrary.wiley.com.proxy.lib.umich.edu/doi/10.1111/1467-842X.00128/abstract%5Cnhttp://onlinelibrary.wiley.com.proxy.lib.umich.edu/store/10.1111/1467-842X.00128/asset/1467-842X.00128.pdf?v=1&t=hp7aamt6&s=734fda260f39b2df398c153d5be84ddd9c7767e2%5Cn},
 	doi = {10.1111/1467-842X.00128},
-	abstract = {This paper describes a technique for computing approximate maximum pseudolikelihood estimates of the parameters of a spatial point process. The method is an extension of Berman \& Turner's (1992) device for maximizing the likelihoods of inhomogeneous spatial Poisson processes. For a very wide class of spatial point process models the likelihood is intractable, while the pseudolikelihood is known explicitly, except for the computation of an integral over the sampling region. Approximation of this integral by a finite sum in a special way yields an approximate pseudolikelihood which is formally equivalent to the (weighted) likelihood of a loglinear model with Poisson responses. This can be maximized using standard statistical software for generalized linear or additive models, provided the conditional intensity of the process takes an ‘exponential family' form. Using this approach a wide variety of spatial point process models of Gibbs type can be fitted rapidly, incorporating spatial trends, interaction between points, dependence on spatial covariates, and mark information.},
 	number = {3},
-	journal = {Australian \& New Zealand Journal of {\textbackslash}textbackslashldots},
+	journal = {Australian \& New Zealand Journal of Statistics},
 	author = {Baddeley, Adrian and Turner, Rolf},
 	year = {2000},
 	pmid = {503},
