+++ NAMESPACE
@@ -41,6 +41,7 @@ importFrom(spatstat.geom,quad)
 importFrom(spatstat.geom,transmat)
 importFrom(stats,complete.cases)
 importFrom(stats,median)
+importFrom(stats,quantile)
 importFrom(stats,runif)
 importFrom(terra,as.polygons)
 importFrom(terra,cellSize)

[…]
