### 2022-12-14 #4771 @skiptoniam
@elbeejay please see the below revisions:

First of all, I would like to thank the reviewers for their thoughtful comments. I have tried to address all their concerns and really appreciate their helpful suggestions.

Responses to @mhesselbarth

I tried to respond to all our open issues for the `ppmData` package. I have closed them all, but please feel free to open them again or create others if you think more needs to be done here.

Some comments considering the manuscript. Overall its written very well and does a good job explaining the package and method. Just some minor suggestions and comments:
Line 8/9: This is an appropriate statistical approach only if locations can be meaningful described by points (e.g., as for plants, nests).
_I have fixed this line and added in a brief example of point pattern data._

Line 13/14: This is one common driver, but also other processes can be more important for the pattern (e.g., competition, neutral theory, etc.). Thus, I would suggest to reword using e.g., “such as”, “for example”, “besides others”.
 _Thanks, that is true. I have added “for example” to make it clear that other factors can determine the observed locations._

Line 15: I would suggest to use locations instead of sightings to be concise.
_Done._

Line 16: Maybe add one more sentence to the previous section to better link habitat conditions to inhomogeneous poisson process models. That should help the reader better to connect the two.
_Added a little bit on environmental conditions._

Line 17-19: I think there is a "that" missing in the sentence.
_Thanks. Fixed._

Line 36-42: Please reword and state only what the package is intended and capable of doing now. Personally, I think the initial intention part is a bit confusing. Also, as far as I understand it, the package can be used for any qualitative marks and not just species. Thus, I would reword the sentence and using species as one example of qualitative marks.
_I think I have reworded this to be more to the point._

Line 66: I understand why the Statement of Need is introduced so late because the reader needs some background about the point processes first. However, I think it might also make sense to move this part up, or at least refer to it earlier in the introduction. But if you decided to keep the structure as it is right now that would be also fine for me.
_We have decided to leave it in this format, as we need to explain what an inhomogeneous point process is before we get into the nitty gritty of estimation._

Line 95-99: Please add some information on what kind of computer/laptop this benchmark is based on (number of cores, RAM, operating system, etc.).
_Done._

Line 107-109: As also mentioned in [Methods to convert from and to spatstat skiptoniam/ppmData#24](https://github.com/skiptoniam/ppmData/issues/24). I would strongly suggest to add conversion methods for spatstat objects or at least explain somewhere in the Vignettes how to convert between spatstat and terra/sf given how popular the spatstat package is.
_I have added a section in vignette that converts a ppmData object to quadrature that can be fitted in spatstat. See lines 182-213 of vignette_

Line 140/141: This sentence could use some more explanation. Why is the step required?
_I have added the reason in the parentheses, packages like spatstat do this step internally._

Responses to @OwenWard

Can you provide a reference for the external data? It might be easier to just include it in the package directly? then it would get a help file, etc
_I have included documentation for external data_

There isn’t an appropriate license for JOSS included
_Fixed this._

The tests which are included don’t currently test any of the methods of the package.
_Fixed this._

There should be a section detailing how to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
_I have included this at the bottom of [readme](https://github.com/skiptoniam/ppmData/tree/joss_submission)_

Given that the statement of need for this paper emphasizes the computational benefits, I think an experiment demonstrating the speedup should be included in the JOSS paper, even commented out, so that readers can confirm it for themselves. I think you could also demonstrate this in the Readme.
_I have added this to the readme._

I'm not sure of the role of the file package.R? It also seems to be somehow linked to the documentation and comes up when you run ?ppmData
_I have updated the package.R file. It now has more detail on the package. This was also one of @mhesselbarth issues_

There are some small typos in the ReadMe in the JOSS branch
In the summary, “Quasi random sampling quadrature are form” and “thus reducing low errors”
_Fixed._

In the example the text mentions red points which are now green
_Fixed._

Also some small typos in the paper on top of those mentioned previously, nothing much
Page 3 line 77, extra “is”
_Fixed. Thanks._

Page 4 line 111-112 maybe missing “is”?
_Fixed._

Page 5 line 130 using “the” glm function maybe?
_Fixed._

The second reference in the bibliography has some missing latex
_Fixed, thanks._
