### 2022-12-15 #4771 @elbeejay
@skiptoniam I do not see where this comment from @OwenWard was addressed:

> There should be a section detailing how to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
I have included this at the bottom of [readme](https://github.com/skiptoniam/ppmData/tree/joss_submission)

I recall having a similar situation during the pre-review phase where some portions of the text I'd flagged were supposed to have been revised, but when I checked they actually were not. I'd like to ask that you double-check all of the files currently on the `joss_submission` branch, to make sure that they reflect the latest revisions and are consistent with the review responses, maybe this is an issue of git branches being out of sync (local v remote).

Additionally, please see the above bot comment which has flagged missing and invalid DOIs, and please correct them as necessary. When that is complete, you can have the bot regenerate both the PDF and the check of the references using the same commands I used above. The bot will automatically refer to the files in the `joss_submission` branch.

Please ping us once those things have been completed, at that point I suspect it will be appropriate to ask our kind reviewers to have another look at this submission when they are able.

Thanks!
