### 2022-12-19 7ebaf74
updates to readme and paper dois
