+++ paper/paper.md
@@ -14,14 +14,14 @@ authors:
     orcid: 0000-0002-6719-8002
     affiliation: "3"
 affiliations:
+ - name: Environment, CSIRO
+   index: 1   
  - name: School of Ecosystems and Forestry Science, The University of Melbourne
-   index: 1
- - name: Oceans & Atmosphere, CSIRO
-   index: 2   
+   index: 2
  - name: Data61, CSIRO
    index: 3    
 citation_author: Woolley & Foster    
-date: 07 July 2022
+date: 14 December 2022
 year: 2022
 bibliography: paper.bib
 header-includes: 
@@ -79,7 +79,7 @@ Another common approach is to use a pseudo-random spatial samples and generate a
 
 Here we use the quasi-random quadrature as a method to trade-off the robustness of pseudo-random sampling and the efficiency of the grid. Quasi-random quadrature spreads the quadrature points in space but simultaneously retains some of the important properties of random sampling. Quasi-random quadrature schemes are available in other R packages, like the excellent `spatstat` [@baddeley_spatial_2015]. The `rQuasi` function can be used for generating a quasi-random spatial sample within a window. However, if the window is not a rectangle, the number of points will be reduced to include only the ones remaining inside the window boundary, so specifying the number of active points is guess work. Our package `ppmData` allows the exact number of quadrature points to be specified within a spatial window.
 
-The package `spatstat` also allows a user to develop a quasi-random quadrature scheme, but this appears to be quite slow when used with Dirichlet weights for larger numbers of dummy (quadrature) points (\> 10000). For example, a regular window with 100000 quadrature points takes approximately 3.5 minutes to run in `spatstat` using `quadscheme` with `method=Dirichlet`. Our package computes the same Dirichlet tessellation in about 4.5 seconds (Comparisons were from a Linux operating system using a Intel i7 8 core processor, with 16 GB of RAM). Typically, calculating a Dirichlet tessellation increases super-linearly and can become very slow for a large numbers of points. Part of the challenge when developing `ppmData` was to make this step efficient by a C++ implementation of a Delaunay triangulation radial sweep algorithm [@sinclair_s-hull_2016]. The Dirichlet tessellations are then calculated as the dual graph of the Delaunay triangulation. Areal weights for the quadrature points are calculated as the area of each Dirichlet tessellation polygon.
+The package `spatstat` also allows a user to develop a quasi-random quadrature scheme, but this appears to be quite slow when used with Dirichlet weights for larger numbers of dummy (quadrature) points (\> 10000). For example, a regular window with 100000 quadrature points takes approximately 2.5 minutes to run in `spatstat` using `quadscheme` with `method=Dirichlet` with quasi-random dummy points. Our package computes the same Dirichlet tessellation in about 7.5 seconds (Comparisons were from a Linux operating system using a Intel i7 8 core processor, with 16 GB of RAM, see package readme for a quick comparison). Typically, calculating a Dirichlet tessellation increases super-linearly and can become very slow for a large numbers of points. Part of the challenge when developing `ppmData` was to make this step efficient by a C++ implementation of a Delaunay triangulation radial sweep algorithm [@sinclair_s-hull_2016]. The Dirichlet tessellations are then calculated as the dual graph of the Delaunay triangulation. Areal weights for the quadrature points are calculated as the area of each Dirichlet tessellation polygon.
 
 ## Generating a quasi-random quadrature scheme
 
