+++ paper/paper.bib
@@ -4,7 +4,6 @@
 	volume = {92},
 	issn = {0012-9658},
 	doi = {10.1890/10-1251.1},
-	abstract = {Community ecologists and conservation biologists often work with data that are too sparse for achieving reliable inference with species-specific approaches. Here we explore the idea of combining species-specific models into a single hierarchical model. The community component of the model seeks for shared patterns in how the species respond to environmental covariates. We illustrate the modeling framework in the context of logistic regression and presence-absence data, but a similar hierarchical structure could also be used in many other types of applications. We first use simulated data to illustrate that the community component can improve parameterization of species-specific models especially for rare species, for which the data would be too sparse to be informative alone. We then apply the community model to real data on 500 diatom species to show that it has much greater predictive power than a collection of independent species-specific models. We use the modeling approach to show that roughly one-third of distance decay in community similarity can be explained by two variables characterizing water quality, rare species typically preferring nutrient-poor waters with high pH, and common species showing a more general pattern of resource use.},
 	number = {2},
 	journal = {Ecology},
 	author = {Ovaskainen, O and Soininen, J},
@@ -33,8 +32,7 @@
 	volume = {190},
 	issn = {03043800},
 	url = {citeulike-article-id:483011%5Cnhttp://dx.doi.org/10.1016/j.ecolmodel.2005.03.026},
-	doi = {doi: 10.1016/j.ecolmodel.2005.03.026},
-	abstract = {The availability of detailed environmental data, together with inexpensive and powerful computers, has fueled a rapid increase in predictive modeling of species environmental requirements and geographic distributions. For some species, detailed presence/absence occurrence data are available, allowing the use of a variety of standard statistical techniques. However, absence data are not available for most species. In this paper, we introduce the use of the maximum entropy method (Maxent) for modeling species geographic distributions with presence-only data. Maxent is a general-purpose machine learning method with a simple and precise mathematical formulation, and it has a number of aspects that make it well-suited for species distribution modeling. In order to investigate the efficacy of the method, here we perform a continental-scale case study using two Neotropical mammals: a lowland species of sloth, Bradypus variegatus, and a small montane murid rodent, Microryzomys minutus. We compared Maxent predictions with those of a commonly used presence-only modeling method, the Genetic Algorithm for Rule-Set Prediction (GARP). We made predictions on 10 random subsets of the occurrence records for both species, and then used the remaining localities for testing. Both algorithms provided reasonable estimates of the species' range, far superior to the shaded outline maps available in field guides. All models were significantly better than random in both binomial tests of omission and receiver operating characteristic (ROC) analyses. The area under the ROC curve (AUC) was almost always higher for Maxent, indicating better discrimination of suitable versus unsuitable areas for the species. The Maxent modeling approach can be used in its present form for many applications with presence-only datasets, and merits further research and development.},
+	doi = {10.1016/j.ecolmodel.2005.03.026},
 	number = {3-4},
 	journal = {Ecological Modelling},
 	author = {Phillips, Steven and Anderson, Robert and Schapire, Robert},
@@ -50,7 +48,6 @@
 	issn = {10510761},
 	url = {%3CGo to},
 	doi = {10.1890/07-2153.1},
[…]
