+++ paper/paper.bib
@@ -227,7 +227,7 @@
 	year = {2013},
 	note = {Publisher: Wiley Online Library},
 	pages = {274--281},
-	doi = {10.1111/j.1541-0420.2012.01824.x.},
+	doi = {10.1111/j.1541-0420.2012.01824.x},
 }
 
 @article{simpson_going_2016,
