### 2022-12-23 #4771 @elbeejay
Thanks @skiptoniam.

@mhesselbarth and @OwenWard it looks like @skiptoniam has worked to address your comments. I'd like to ask that you finish your reviews when you are able - maybe we aim for the end of January? Let me know if that is not going to be possible and we can work to set a reasonable timeline. Thank you again for reviewing for JOSS. Please ping me with any questions about the review process.
