### 2023-01-31 skiptoniam/ppmData#30 @OwenWard
Small typo in documentation
Very minor, but I spotted a typo in the documentation

In the details for ppmData, I think you mean "100000 random numbers" rather than "100000 halton numbers"
