### 2023-01-31 #4771 @OwenWard
I've had another look through the package and everything looks good. I marked two small issues but they're easy to fix.

- I'm still a little confused about the need for the `package.R` file?
- For the comparison in the README, I'm likely confused but is there any reason you don't use the `snails` data in the comparison with `spatstat`? Basically, why not make the comparisons as exact as possible?

Once those and the 2 issues are addressed I'm happy @elbeejay
