### 2023-02-17 skiptoniam/ppmData#32 @mhesselbarth
Test for class
Regarding https://github.com/openjournals/joss-reviews/issues/4771

There still are two class(x) tests left in `dirtess.R` and `extras.R`.
