### 2023-02-17 #4771 @mhesselbarth
Okay, again sorry for the delay! I finally had the chance to look at the package again. I pretty much agree with @OwenWard.

1. There still are two warnings thrown due to `class(x)` checks (instead of `inherits()`).
https://github.com/skiptoniam/ppmData/issues/32

2. I think the identical name of the package and the main function is still confusing to me. Interstingly, this also combines the documention from `package.R` and `ppmData.R` when using `?ppmData`. I don't think this is an actual problem, but it definitely feels a bit strange and I'm not sure if this could introduce issues down to road. So, I would suggest to change probably the name of the main function e.g., `fit_ppmData()`
https://github.com/skiptoniam/ppmData/issues/22
