### 2023-02-22 bc2308a
update to readme, package file and class warning in checks
