+++ README.Rmd
@@ -84,12 +84,14 @@ system.time(ppmdata2 <- ppmData(npoints = 100000,
                                 covariates=covariates))
 ```
 
-Using a simple spatstat example to generate a quasi-random scheme
+Using a spatstat to generate a quasi-random scheme for the snails dataset.
 ```{r}
 suppressPackageStartupMessages(library(spatstat))
-data(simdat)
-D <- rQuasi(100000,Window(simdat))
-system.time(Q <- quadscheme(simdat,D,method="dirichlet", exact=FALSE))
+e <- ext(covariates[[1]])
+W <- owin(e[1:2],e[3:4])
+snails_ppp <- ppp(presences[,1],presences[,2],W)
+D <- rQuasi(100000,W)
+system.time(Q <- quadscheme(snails_ppp,D,method="dirichlet", exact=FALSE))
 ```
 
 
