+++ man/ppmData.Rd
@@ -1,6 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/package.R, R/ppm_data.R
-\docType{package}
+% Please edit documentation in R/ppm_data.R
 \name{ppmData}
 \alias{ppmData}
 \title{Develop a quadrature scheme using quasi-random sampling for spatial
@@ -90,18 +89,6 @@ thus reducing low errors for geostatistical prediction (Diggle & Ribeiro,
 2007). Because the quasi-random design is not on a regular grid we use
 Dirichlet tessellation to generate polygons for each point in the quadrature
 scheme. Areal weights are then derived from these polygons.
-
-This package is a way to efficiently generate a quasi-random set
-of background points for presence-only modelling of single or multiple
-responses. The package was set up to model multiple species presence-only
-data sets, but could be used for an spatial point process modelling.
-Quasi-random points are a nice alternative to pseudo-random samples, this is
-because we can generate a quasi-random sample across and areal region
-(X and Y coordinates). This in turn should reduce autocorrelation in
-quadrature scheme. The weight of each quadrature point is calculated using
-Dirichlet (Voronoi) Tessellation written in c++. We calculated the duel-graph
-of a Delaunay triangulation. The Delaunay triangulation is constructed based
-on a radial sweep algorithm.
 }
 \details{
 The approach uses quasi-random sampling to generate a quadrature
