+++ DESCRIPTION
@@ -8,7 +8,7 @@ Maintainer: Skipton Woolley <skip.woolley@csiro.au>
 Description: This package generates a set of quasi-random points for a quadrature scheme, we calculate each points weights using a dirichlet tesselation (voronoi) algorithm. The resulting object can be used in existing single and multiple species models in order to extend them to Point Processes.
 License: MIT + file LICENSE
 LazyData: TRUE
-RoxygenNote: 7.2.0
+RoxygenNote: 7.2.3
 URL: https://github.com/skiptoniam/ppmData
 BugReports: https://github.com/skiptoniam/ppmData/issues
 Imports: 
