+++ README.md
@@ -123,7 +123,7 @@ system.time(ppmdata2 <- ppmData(npoints = 100000,
     ## There are a total of 100045 sites in the model.matrix
 
     ##    user  system elapsed 
-    ##   4.348   0.008   4.360
+    ##   4.355   0.024   4.382
 
 Using a spatstat to generate a quasi-random scheme for the snails
 dataset.
@@ -133,17 +133,12 @@ suppressPackageStartupMessages(library(spatstat))
 e <- ext(covariates[[1]])
 W <- owin(e[1:2],e[3:4])
 snails_ppp <- ppp(presences[,1],presences[,2],W)
-```
-
-    ## Warning: data contain duplicated points
-
-``` r
 D <- rQuasi(100000,W)
 system.time(Q <- quadscheme(snails_ppp,D,method="dirichlet", exact=FALSE))
 ```
 
     ##    user  system elapsed 
-    ## 266.514   0.484 267.097
+    ## 258.793   0.507 259.491
 
 ## Code of Conduct
 
