### 2023-02-22 #4771 @skiptoniam
Hi All,

I have update the comparison in the README as per @OwenWard first comment. I also think that this is a good idea and have updated the comparison to be run on the snails data for both bits of software.

I have removed the package.R file and moved any additional information (in the package.R file) into the ppmData.R file. I think this will deal with any conflict and issues associated with the package file. And should address point 1 from @OwenWard and point 2 from @mhesselbarth.

Apologies on missing that class(x) call as mentioned by @mhesselbarth, I thought I had caught all of those previously. I just ran the checks as part of the package build and they all appear to be dealt with now.

Thanks again to @elbeejay, @OwenWard & @mhesselbarth for all your really helpful suggestions and comments.
