### 2023-02-23 #4771 @elbeejay
@skiptoniam a few minor things I'd like to close out before we move on to getting metadata set and everything ahead of publication.

1. As best I can tell, issue 30: https://github.com/skiptoniam/ppmData/issues/30 remains unresolved
2. It'd also be preferable (from the JOSS perspective) if you could, at the minimum, address, if not resolve, the other two outstanding issues in the repository that were raised as part of this review process

Thanks
