### 2023-02-24 #4771 @skiptoniam
@elbeejay I have fixed those issues and closed them.

Cheers,

Skip.
