### 2023-02-24 #4771 @elbeejay
Great, thanks @skiptoniam. At this time we need to finalize the metadata and versioning for the version of `ppmData` that will be tied to this JOSS publication. I noticed that the work related to the JOSS publication has happened in the "joss_submission" branch, this is also an opportunity to merge it into the "main" branch if you'd like, as the review process is over. Please let me know once you've been able to do the following:

- [ ] (if necessary) Re-release the package on CRAN so that the "JOSS version" matches one in a package repository

Thanks again to @OwenWard and @mhesselbarth for their work reviewing this package.
