### 2023-02-27 0c12183
update to package - fix vignette error
