### 2023-02-27 #4771 @skiptoniam
@elbeejay Thanks.

I have made a release: https://github.com/skiptoniam/ppmData/releases/tag/v1.0.0
Archived the software here: https://doi.org/10.5281/zenodo.7679406
I am yet to put the software on CRAN, but will do so soon.
