### 2023-02-27 #4771 @elbeejay
@skiptoniam thank you for doing that. A release on CRAN is not a requirement for JOSS, just a suggestion and a reminder I like for include as some R packages submitted to JOSS are already hosted on JOSS (and could be updated to match the JOSS version of record).

At this time I will be recommending ppmData for publication in JOSS. Sometime over the next week or so an editor-in-chief will review the paper and this issue to do a final check before officially publishing your article in JOSS.
